Failed attempt to reduce duplicate code
I attempted to reuse the same file upload script in order to administrators to upload images in the update page via a GET parameter... this didn't work because PHP is assuming there's a file with the literal name of "fileupload.php?id=". I might redo this with a Header instead of an includes statement and see if that works.