Image Updating is now fixed.
Finally fixed it, I ended up having to take all of the file upload code and pushing it to its own function inside the process_post script. This works for pages with existing pictures as well as pages without images.

I also ran into another issue just before this. When I attempted to query the database for an existing image row, it threw a null reference exception saying that my $db object was null. This was an easy fix as all I had to do was pass the $db as an argument in the updateImages function.