Deleting Uploaded Images
Functionality has now been included to delete any images currently uploaded. Images will be deleted from both the uploads folder and the images table.

If a table doesn't have an image associated with it, the administrator won't be prompted with a "delete image" checkbox. Images can also be deleted when updating the page, and will automatically be deleted when deleting pages.